requirements.txt
bottle
Pillow
pdf2image
beaker
yomitoku
opencv-python
torch